Move all contents
