+++ static/app/views/insights/browser/resources/components/renderBlockingSelector.tsx
@@ -3,10 +3,10 @@ import {trackAnalytics} from 'sentry/utils/analytics';
 import {browserHistory} from 'sentry/utils/browserHistory';
 import {useLocation} from 'sentry/utils/useLocation';
 import useOrganization from 'sentry/utils/useOrganization';
+import type {Option} from 'sentry/views/insights/browser/resources/components/selectControlWithProps';
+import SelectControlWithProps from 'sentry/views/insights/browser/resources/components/selectControlWithProps';
 import {QueryParameterNames} from 'sentry/views/insights/common/views/queryParameters';
 import {SpanMetricsField} from 'sentry/views/insights/types';
-import type {Option} from 'sentry/views/performance/browser/resources/shared/selectControlWithProps';
-import SelectControlWithProps from 'sentry/views/performance/browser/resources/shared/selectControlWithProps';
 
 const {RESOURCE_RENDER_BLOCKING_STATUS} = SpanMetricsField;
