+++ static/app/views/insights/browser/resources/components/resourceSize.tsx
