+++ static/app/views/insights/browser/resources/components/sampleImages.spec.tsx
@@ -7,8 +7,8 @@ import ProjectsStore from 'sentry/stores/projectsStore';
 import type {Organization} from 'sentry/types';
 import {useLocation} from 'sentry/utils/useLocation';
 import usePageFilters from 'sentry/utils/usePageFilters';
+import SampleImages from 'sentry/views/insights/browser/resources/components/sampleImages';
 import {SpanIndexedField} from 'sentry/views/insights/types';
-import SampleImages from 'sentry/views/performance/browser/resources/resourceSummaryPage/sampleImages';
 
 const {SPAN_GROUP, HTTP_RESPONSE_CONTENT_LENGTH, RAW_DOMAIN, SPAN_DESCRIPTION} =
   SpanIndexedField;
