+++ static/app/views/insights/browser/resources/components/tables/resourceTable.tsx
@@ -15,6 +15,17 @@ import {DismissId, usePageAlert} from 'sentry/utils/performance/contexts/pageAle
 import {decodeScalar} from 'sentry/utils/queryString';
 import {useLocation} from 'sentry/utils/useLocation';
 import useOrganization from 'sentry/utils/useOrganization';
+import {useResourcesQuery} from 'sentry/views/insights/browser/common/queries/useResourcesQuery';
+import {
+  FONT_FILE_EXTENSIONS,
+  IMAGE_FILE_EXTENSIONS,
+} from 'sentry/views/insights/browser/resources/constants';
+import {
+  DATA_TYPE,
+  RESOURCE_THROUGHPUT_UNIT,
+} from 'sentry/views/insights/browser/resources/settings';
+import {ResourceSpanOps} from 'sentry/views/insights/browser/resources/types';
+import type {ValidSort} from 'sentry/views/insights/browser/resources/utils/useResourceSort';
 import {DurationCell} from 'sentry/views/insights/common/components/tableCells/durationCell';
 import {renderHeadCell} from 'sentry/views/insights/common/components/tableCells/renderHeadCell';
 import ResourceSizeCell from 'sentry/views/insights/common/components/tableCells/resourceSizeCell';
@@ -27,15 +38,6 @@ import {
   getThroughputTitle,
 } from 'sentry/views/insights/common/views/spans/types';
 import {ModuleName, SpanFunction, SpanMetricsField} from 'sentry/views/insights/types';
-import {RESOURCE_THROUGHPUT_UNIT} from 'sentry/views/performance/browser/resources';
-import {DATA_TYPE} from 'sentry/views/performance/browser/resources/settings';
-import {
-  FONT_FILE_EXTENSIONS,
-  IMAGE_FILE_EXTENSIONS,
-} from 'sentry/views/performance/browser/resources/shared/constants';
-import {ResourceSpanOps} from 'sentry/views/performance/browser/resources/shared/types';
-import type {ValidSort} from 'sentry/views/performance/browser/resources/utils/useResourceSort';
-import {useResourcesQuery} from 'sentry/views/performance/browser/resources/utils/useResourcesQuery';
 
 const {
   SPAN_DESCRIPTION,
