+++ static/app/views/insights/browser/resources/constants.ts
