+++ static/app/views/insights/browser/resources/queries/useIndexedResourceQuery.ts
