+++ static/app/views/insights/browser/resources/queries/useResourcePageQuery.ts
