+++ static/app/views/insights/browser/resources/queries/useResourcePagesQuery.ts
@@ -4,13 +4,13 @@ import {DiscoverDatasets} from 'sentry/utils/discover/types';
 import {useLocation} from 'sentry/utils/useLocation';
 import useOrganization from 'sentry/utils/useOrganization';
 import usePageFilters from 'sentry/utils/usePageFilters';
-import {SpanMetricsField} from 'sentry/views/insights/types';
-import {useResourceModuleFilters} from 'sentry/views/performance/browser/resources/utils/useResourceFilters';
 import {
   DEFAULT_RESOURCE_FILTERS,
   getDomainFilter,
   getResourceTypeFilter,
-} from 'sentry/views/performance/browser/resources/utils/useResourcesQuery';
+} from 'sentry/views/insights/browser/common/queries/useResourcesQuery';
+import {useResourceModuleFilters} from 'sentry/views/insights/browser/resources/utils/useResourceFilters';
+import {SpanMetricsField} from 'sentry/views/insights/types';
 
 const {SPAN_DOMAIN, SPAN_OP} = SpanMetricsField;
