+++ static/app/views/insights/browser/resources/referrer.ts
