+++ static/app/views/insights/browser/resources/types.tsx
