+++ static/app/views/insights/browser/resources/utils/useResourceSummarySort.ts
