+++ static/app/views/insights/browser/resources/views/resourceSummaryPage.tsx
@@ -13,20 +13,20 @@ import {t, tct} from 'sentry/locale';
 import {MutableSearch} from 'sentry/utils/tokenizeSearch';
 import {useLocation} from 'sentry/utils/useLocation';
 import {useParams} from 'sentry/utils/useParams';
+import ResourceSummaryCharts from 'sentry/views/insights/browser/resources/components/charts/resourceSummaryCharts';
+import RenderBlockingSelector from 'sentry/views/insights/browser/resources/components/renderBlockingSelector';
+import ResourceInfo from 'sentry/views/insights/browser/resources/components/resourceInfo';
+import {FilterOptionsContainer} from 'sentry/views/insights/browser/resources/components/resourceView';
+import SampleImages from 'sentry/views/insights/browser/resources/components/sampleImages';
+import ResourceSummaryTable from 'sentry/views/insights/browser/resources/components/tables/resourceSummaryTable';
+import {IMAGE_FILE_EXTENSIONS} from 'sentry/views/insights/browser/resources/constants';
+import {Referrer} from 'sentry/views/insights/browser/resources/referrer';
+import {DATA_TYPE} from 'sentry/views/insights/browser/resources/settings';
+import {ResourceSpanOps} from 'sentry/views/insights/browser/resources/types';
+import {useResourceModuleFilters} from 'sentry/views/insights/browser/resources/utils/useResourceFilters';
 import {useSpanMetrics} from 'sentry/views/insights/common/queries/useDiscover';
 import {SampleList} from 'sentry/views/insights/common/views/spanSummaryPage/sampleList';
 import {ModuleName, SpanMetricsField} from 'sentry/views/insights/types';
-import {Referrer} from 'sentry/views/performance/browser/resources/referrer';
-import ResourceInfo from 'sentry/views/performance/browser/resources/resourceSummaryPage/resourceInfo';
-import ResourceSummaryCharts from 'sentry/views/performance/browser/resources/resourceSummaryPage/resourceSummaryCharts';
-import ResourceSummaryTable from 'sentry/views/performance/browser/resources/resourceSummaryPage/resourceSummaryTable';
-import SampleImages from 'sentry/views/performance/browser/resources/resourceSummaryPage/sampleImages';
-import {FilterOptionsContainer} from 'sentry/views/performance/browser/resources/resourceView';
-import {DATA_TYPE} from 'sentry/views/performance/browser/resources/settings';
-import {IMAGE_FILE_EXTENSIONS} from 'sentry/views/performance/browser/resources/shared/constants';
-import RenderBlockingSelector from 'sentry/views/performance/browser/resources/shared/renderBlockingSelector';
-import {ResourceSpanOps} from 'sentry/views/performance/browser/resources/shared/types';
-import {useResourceModuleFilters} from 'sentry/views/performance/browser/resources/utils/useResourceFilters';
 import {ModulePageProviders} from 'sentry/views/performance/modulePageProviders';
 import {TraceViewSources} from 'sentry/views/performance/newTraceDetails/traceMetadataHeader';
 import {useModuleBreadcrumbs} from 'sentry/views/performance/utils/useModuleBreadcrumbs';
