+++ static/app/views/insights/browser/resources/views/resourcesLandingPage.tsx
@@ -11,30 +11,27 @@ import PageFilterBar from 'sentry/components/organizations/pageFilterBar';
 import {ProjectPageFilter} from 'sentry/components/organizations/projectPageFilter';
 import {PageHeadingQuestionTooltip} from 'sentry/components/pageHeadingQuestionTooltip';
 import {space} from 'sentry/styles/space';
-import {RateUnit} from 'sentry/utils/discover/fields';
 import {PageAlert, PageAlertProvider} from 'sentry/utils/performance/contexts/pageAlert';
-import {DomainSelector} from 'sentry/views/insights/common/views/spans/selectors/domainSelector';
-import {ModuleName} from 'sentry/views/insights/types';
+import {DEFAULT_RESOURCE_FILTERS} from 'sentry/views/insights/browser/common/queries/useResourcesQuery';
 import ResourceView, {
-  DEFAULT_RESOURCE_TYPES,
   FilterOptionsContainer,
-} from 'sentry/views/performance/browser/resources/resourceView';
+} from 'sentry/views/insights/browser/resources/components/resourceView';
 import {
+  DEFAULT_RESOURCE_TYPES,
   MODULE_DESCRIPTION,
   MODULE_DOC_LINK,
-} from 'sentry/views/performance/browser/resources/settings';
+} from 'sentry/views/insights/browser/resources/settings';
 import {
   BrowserStarfishFields,
   useResourceModuleFilters,
-} from 'sentry/views/performance/browser/resources/utils/useResourceFilters';
-import {DEFAULT_RESOURCE_FILTERS} from 'sentry/views/performance/browser/resources/utils/useResourcesQuery';
+} from 'sentry/views/insights/browser/resources/utils/useResourceFilters';
+import {DomainSelector} from 'sentry/views/insights/common/views/spans/selectors/domainSelector';
+import {ModuleName} from 'sentry/views/insights/types';
 import {ModulePageProviders} from 'sentry/views/performance/modulePageProviders';
 import {useModuleBreadcrumbs} from 'sentry/views/performance/utils/useModuleBreadcrumbs';
 import {useModuleTitle} from 'sentry/views/performance/utils/useModuleTitle';
 
 const {SPAN_OP, SPAN_DOMAIN} = BrowserStarfishFields;
-
-export const RESOURCE_THROUGHPUT_UNIT = RateUnit.PER_MINUTE;
 
 function ResourcesLandingPage() {
   const filters = useResourceModuleFilters();
