+++ static/app/views/insights/browser/webVitals/components/recommendations.tsx
@@ -5,9 +5,9 @@ import {IconFile} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
 import getDuration from 'sentry/utils/duration/getDuration';
+import {useResourcesQuery} from 'sentry/views/insights/browser/common/queries/useResourcesQuery';
 import type {WebVitals} from 'sentry/views/insights/browser/webVitals/types';
 import {SpanMetricsField} from 'sentry/views/insights/types';
-import {useResourcesQuery} from 'sentry/views/performance/browser/resources/utils/useResourcesQuery';
 
 export function Recommendations({
   transaction,
